+++ src/Microsoft.Azure.Functions.Worker.Extensions.Mcp.Sdk/FunctionsMiddleware/FunctionsMcpToolResultMiddleware.cs
@@ -68,6 +68,7 @@ internal class FunctionsMcpToolResultMiddleware : IFunctionsWorkerMiddleware
                 // Don't process CallToolResult - just serialize as-is
                 type = Constants.CallToolResultType;
                 content = JsonSerializer.Serialize(callToolResult, McpJsonUtilities.DefaultOptions);
+                structuredContent = callToolResult.StructuredContent?.ToJsonString();
                 break;
 
             case ContentBlock block: